pyproject.toml
[project]
name = "pilot-decoder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "hydra-core>=1.3.2",
    "mlflow>=3.9.0",
    "torch>=2.6.0",
]
